Hoist the style parts of sequential.rs into style/.
